+++ CHANGELOG.md
@@ -5,6 +5,17 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+## [1.0.0-alpha.19] - 2019-05-05
+
+### Changed
+
+- Reduce frequency to call `requestAnimationFrame`
+
+### Added
+
+- `mount()` now returns fragment with `mount` method
+
+
 ## [1.0.0-alpha.18] - 2019-05-04
 
 ### Fixed